AccessRight: split code + integrated feedback
